# Improved the performance of TableServiceClient.GetTableClient()