| Method          | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|---------------- |----------:|----------:|----------:|------:|--------:|-------:|-------:|----------:|------------:|
| GetTableClient_New |  1.544 μs | 0.0303 μs | 0.0298 μs |  0.03 |    0.00 | 0.1812 |      - |   1.49 KB |        0.06 |
| GetTableClient  | 57.302 μs | 0.7423 μs | 0.6580 μs |  1.00 |    0.02 | 3.0518 | 0.1831 |  25.21 KB |        1.00 |